+++ tests/system/detail/test_attribute_table.cc
@@ -73,3 +73,42 @@ TEST_CASE("attribute_table::require - is idempotent")
     CHECK(values[2] == 42);
     CHECK(values[3] == 42);
 }
+
+TEST_CASE("attribute_table - is copyable")
+{
+    md::detail::attribute_table table;
+
+    md::attribute_key<int, struct test> test_attribute = [](test*) -> int {
+        return 42;
+    };
+
+    table.require(test_attribute);
+    table.resize(4);
+
+    SECTION("copy construction")
+    {
+        md::detail::attribute_table copy = table;
+        CHECK(copy.size() == 4);
+
+        md::array_view<int> values = copy.view(test_attribute);
+        CHECK(values.size() == 4);
+        CHECK(values[0] == 42);
+        CHECK(values[1] == 42);
+        CHECK(values[2] == 42);
+        CHECK(values[3] == 42);
+    }
+
+    SECTION("copy assignment")
+    {
+        md::detail::attribute_table copy;
+        copy = table;
+        CHECK(copy.size() == 4);
+
+        md::array_view<int> values = copy.view(test_attribute);
+        CHECK(values.size() == 4);
+        CHECK(values[0] == 42);
+        CHECK(values[1] == 42);
+        CHECK(values[2] == 42);
+        CHECK(values[3] == 42);
+    }
+}